requirements.txt
openai==0.28
PyMuPDF
fuzzywuzzy
Pillow
fastapi 
uvicorn 
mangum
python-Levenshtein
